Allow to use custom serializer with primitive type

+++ src/JMS/Serializer/SerializationContext.php
@@ -48,12 +48,18 @@ class SerializationContext extends Context
 
     public function startVisiting($object)
     {
+        if ( ! is_object($object)) {
+            return;
+        }
         $this->visitingSet->attach($object);
         $this->visitingStack->push($object);
     }
 
     public function stopVisiting($object)
     {
+        if ( ! is_object($object)) {
+            return;
+        }
         $this->visitingSet->detach($object);
         $poppedObject = $this->visitingStack->pop();
 
@@ -65,7 +71,7 @@ class SerializationContext extends Context
     public function isVisiting($object)
     {
         if ( ! is_object($object)) {
-            throw new LogicException('Expected object but got '.gettype($object).'. Do you have the wrong @Type mapping or could this be a Doctrine many-to-many relation?');
+            return false;
         }
 
         return $this->visitingSet->contains($object);
